bench/pyproject.toml
[project]
name = "bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "turbopuffer>=1.0.0",
    "openai>=1.0.0",
    "beir>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]
